% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstats.R
\name{dstats}
\alias{dstats}
\title{Descriptive Statistics}
\usage{
dstats(fit)
}
\arguments{
\item{fit}{a regression object.}
}
\value{
A \code{pstools_dstats} object which is a
\code{data.frame}, with descriptive statistics for each variable in and out
of the exposure group.
}
\description{
Given a logistic regression model, generate the propensity weights for the
subjects (rows) in the data set and report the descriptive statistics for
each predictor in the regression model.
}
\details{
The regression model is expected to estimate the probability of an exposure
(Z = 1) given a set of predictors, X, i.e., Pr[Z = 1 | X].

NOTE: for binary predictors coded as 0/1, such as male, the default action of
\code{dstats} will return a mean (sd), that is, if the \code{formula} for the
regression model is of the form
\code{y ~ male}, \code{dstats} will summarize the variable \code{male} as if it
was continous predictor. To get the summary of the proportion of males, and
females will be reported too, use a regression formula of the form \code{y ~
factor(male)}.

See \code{\link{psweights}} for details on the propensity score based weights
used by \code{dstats}.
}
\examples{

data(pride)
glmfit <- stats::glm(PCR_RSV ~ SEX + RSVINF + REGION + AGE + ELTATOP + EINZ + EXT, 
                     data = pride,
                     family = stats::binomial())
dstats(glmfit)

}
\references{
Li, Liang, and Tom Greene. "A weighting analogue to pair matching in
propensity score analysis." The international journal of biostatistics 9.2
(2013): 215-234.
}
\seealso{
\code{\link[stats]{glm}} for fitting logistic regression models, or 
\code{\link[geepack]{geeglm}} for fitting GEEs.
\code{\link{psweights}} for the weights.
}

