% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_stepper.R
\name{gee_stepper}
\alias{gee_stepper}
\title{Stepwise GEE Model Selection}
\usage{
gee_stepper(fit, upper)
}
\arguments{
\item{fit}{a \code{geeglm} object.  the response, id, data, corstr, ... will
be reused in all the the fits called by \code{gee_stepper}}

\item{upper}{A formula with the rhs containing all the variables that *could*
be the the model.}
}
\value{
the "best" model based on QIC from the stepwise selection process.
}
\description{
Starting with a model of y ~ 1, use a forward/backward step selection process
to find the preferable fit.
}
\examples{

data(pride)
library(geepack)
geefit <- geepack::geeglm(PCR_RSV ~ SEX + RSVINF + REGION + AGE + ELTATOP + EINZ + EXT, 
                  id = REGION, data = pride, family = stats::binomial())
gee_stepper(geefit, formula(geefit))

}

