% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardized_differences.R
\name{standardized_differences}
\alias{standardized_differences}
\title{Standardized Differences}
\usage{
standardized_differences(x, ...)
}
\arguments{
\item{x}{a logistic regression object, or a \code{pstools_dstats} object.}

\item{...}{ignored at this time.}
}
\value{
A list of \code{data.frame}s each element of the list is for a
different weighting method.
}
\description{
Calculate the standardized differences (percentages).
}
\examples{
data(pride)
glmfit <- stats::glm(PCR_RSV ~ SEX + RSVINF + REGION + AGE + ELTATOP + EINZ + EXT, 
                     data = pride,
                     family = stats::binomial())

standardized_differences(glmfit)
}
\seealso{
\code{\link{psweights}}, \code{\link{dstats}}
}

