% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psweights.R
\name{psweights}
\alias{psweights}
\alias{psweights_}
\title{Propensity Score Weights}
\usage{
psweights(.data, exposure, ps)

psweights_(.data, exposure_col, ps_col)
}
\arguments{
\item{.data}{a \code{data.frame}}

\item{exposure}{the bare name for the column within \code{.data} indicating
the exposure or non-exposure.  Expect this to be an integer, or coercible to
an integer with values of zero or one.}

\item{ps}{the propensity scores.  Expected values between 0 and 1.}

\item{exposure_col}{a character string}

\item{ps_col}{a character string}
}
\value{
a \code{data.frame} with the exposure and ps vectors returned along
with five different weights.  See Details for information on the five
weights.
}
\description{
Five different matching weights based on an exposure and propensity score.
}
\details{
\tabular{ll}{
Average causal effect in study population                                \tab \code{psw_IPW}           \cr
Average causal effect in exposed                                         \tab \code{psw_ACE_Exposed}   \cr
Average causal effect in unexposed                                       \tab \code{psw_ACE_Unexposed} \cr
Average causes effect in population for which sample is most informative \tab \code{psw_ACE_MostInfo}  \cr
Average causal effect in mathcin weight population                       \tab \code{psw_ACE_MWP}       \cr
}
}
\examples{
glmfit <- stats::glm(PCR_RSV ~ SEX + RSVINF + REGION + AGE + ELTATOP + EINZ + EXT, 
                     data = pride,
                     family = stats::binomial()) 

ourpride <- dplyr::mutate(pride, pp = qwraps2::invlogit(fitted(glmfit))) 

psweights(dplyr::tbl_df(ourpride), PCR_RSV, pp)

}
\references{
Li, Liang, and Tom Greene. "A weighting analogue to pair matching in
propensity score analysis." The international journal of biostatistics 9.2
(2013): 215-234.
}

