% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardized_differences.R
\name{standardized_differences}
\alias{plot.pstools_std_diffs}
\alias{standardized_differences}
\title{Standardized Differences}
\usage{
standardized_differences(x, ...)

\method{plot}{pstools_std_diffs}(x, adj_methods = c("adj_IPW",
  "adj_ACE_Exposed", "adj_ACE_Unexposed", "adj_ACE_MostInfo", "adj_ACE_MWP"),
  ...)
}
\arguments{
\item{x}{a logistic regression object, or a \code{pstools_dstats} object.}

\item{...}{ignored at this time.}

\item{adj_methods}{a character vector, names standardized differences
adjustment methods to include in the plot.}
}
\value{
A list of \code{data.frame}s each element of the list is for a
different weighting method.
}
\description{
Calculate the standardized differences (percentages).
}
\section{Methods (by class)}{
\itemize{
\item \code{pstools_std_diffs}: plotting method for
\code{pstools_std_diffs} objects.
}}
\examples{
data(pride)
glmfit <- stats::glm(PCR_RSV ~ SEX + RSVINF + REGION + AGE + ELTATOP + EINZ + EXT, 
                     data = pride,
                     family = stats::binomial())

std_diff <- standardized_differences(glmfit)

std_diff

# the plot is a ggplot2 object
plot(std_diff)

plot(std_diff, adj_methods = "adj_ACE_MWP") + 
  ggplot2::ggtitle("ACE MWP vs Unadjusted")
}
\seealso{
\code{\link{psweights}}, \code{\link{dstats}}
}

